+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Fixed
 - Bad config merging in ``handle_subcommands`` pytorch-lightning#10859
   <https://github.com/PyTorchLightning/pytorch-lightning/issues/10859>`__.
 - Unit tests failing with argcomplete>=2.0.0.
+- Namespace TypeError with non-str inputs `#116
+  <https://github.com/omni-us/jsonargparse/issues/116>`__.
 
 
 v4.1.0 (2021-12-06)